Update index.html
Created 6 blocks of 10 trials and instead of hardcoding the fractions, they are now randomized.

The values are randomized to be roughly 50% matches and 50% non-matches, but this is a very rough randomization.

There is no longer feedback for the experiment and only feedback in the example trial.

Also an update to the look of the fraction